Segunda Iteración

+++ src/main/java/ar/com/ddsutn/TP1Macowins/Camisa.java
@@ -3,7 +3,7 @@ package ar.com.ddsutn.TP1Macowins;
 public class Camisa extends Prenda{
 	
 	@Override
-	protected int precioBase(){
+	protected double precioBase(){
 		return 200;
 	}
